Create GA_bugy.py
buggy script:)
i need to put a check on like line 394 bc some jobs are being submitted but only 1 run is being executed so instead of the 3 its only 1 which is a problem
so im thinking maybe i put a check to make sure the 3 are being submitted and if the job is still in the queue then thats fine but if its gone then i resubmit it? thoughts?